+++ apps/devtools-extension/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @blac/devtools-extension
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @blac/devtools-ui@2.0.1
+  - @blac/react@2.0.1
+  - @blac/core@2.0.1
+
 ## 2.0.0
 
 ### Patch Changes