+++ apps/examples/CHANGELOG.md
@@ -1,5 +1,16 @@
 # examples
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @blac/devtools-connect@2.0.1
+  - @blac/logging-plugin@2.0.1
+  - @blac/devtools-ui@2.0.1
+  - @blac/react@2.0.1
+  - @blac/core@2.0.1
+
 ## 1.0.1
 
 ### Patch Changes